pyproject.toml
[tool.poetry]
name = "wzlight"
version = "0.1.1"
description = "Light asynchronous wrapper for COD Warzone API"
authors = ["Matthieu Vion <viomatthieu@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/matthieuvion/wzlight"
keywords = ["warzone", "API", "asynchronous", "httpx"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.23.0"
python-dotenv = "^0.20.0"
backoff = "^2.1.2"
async-lru = "^1.0.3"
jupyterlab = "^3.4.3"

[tool.poetry.dev-dependencies]
black = {version = "^22.3.0", allow-prereleases = true}
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
